@@ -6,17 +6,25 @@ class AttachmentsController < ApplicationController
     @attachments = Attachment.where(public: 1).limit(50)
     respond_to do |format|
       format.json { 
-        render json: @attachments.collect{|a| a.to_jq_upload }.to_json 
+        render json: @attachments
       }
     end
   end
 
   def create
-    @attachment = Attachment.new(:attachment => params[:attachments])
+    attachment_params = {}
+    attachment_params[:attachment] = params[:attachments] || params[:photo][:imagefile]
+    attachment_params[:mime]       = params[:mime]        || params[:photo][:imagefile].content_type
+    attachment_params[:name]       = params[:name]        || params[:photo][:imagefile].original_filename
+    @attachment = Attachment.new(attachment_params)
     respond_to do |format|
       if @attachment.save
+        format.html {
+          redirect_to @attachment, notice: 'Attachment successfully created'
+        }
         format.json { 
-          render json: [@attachment.to_jq_upload].to_json,
+          data = { id: @attachment.id, thumb: view_context.image_tag(@attachment.attachment.url(:thumb)) }
+          render json: data,
                  status: :created, 
                  location: @attachment 
         }
@@ -31,8 +39,9 @@ class AttachmentsController < ApplicationController
 
   def update
     @attachment = Attachment.find(params[:id])
+    attachment_attributes = params[:attachments] || params[:photo][:imagefile]
     respond_to do |format|
-      if @attachment.update_attributes(params[:attachments])
+      if @attachment.update_attributes(attachment_attributes)
         format.json { head :no_content }
       else 
         format.json { render json: @attachment.errors, status: :unprocessable_entity }
