@@ -2,7 +2,7 @@ class SearchController < ApplicationController
   def index
     @results = []
     if params[:q]
-      @results = Post.fast_search(params[:q])
+      @results = Post.search(params[:q])
       pp @results
       respond_to do |format|
         format.html { render action: "index" }
