@@ -1,9 +1,6 @@
 class Category < ActiveRecord::Base
-  attr_accessible :name, :id
-  has_many :posts
   extend Hashifiable
   hashify :name, :id
 
-  validates :name, presence: true
-  validates :name, uniqueness: true
+  validates :name, presence: true, uniqueness: true
 end
